+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -54,6 +54,9 @@
     "name": "BrowserModule"
   },
   {
+    "name": "BrowserXhr"
+  },
+  {
     "name": "BuiltInControlValueAccessor"
   },
   {
