+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -81,6 +81,9 @@
     "name": "BrowserViewportScroller"
   },
   {
+    "name": "BrowserXhr"
+  },
+  {
     "name": "CIRCULAR"
   },
   {
